refactor config for easier import
